requirements.txt
pandas==2.2.3
pyyaml==6.0.2
requests==2.32.3
python-dotenv==1.1.0
numpy==2.2.6
tqdm==4.66.2 